Add files via upload
Final Project description: 
Milestone 1: deliver a first design of the data analysis pipeline, and identify the two external ontologies and two datasets that you will use.
Milestone 2: "deliver a conceptualization and representation of the ontology that your analysis will use, by reusing the two ontologies you identified in the first milestone. In addition, you have to create the instances by integrating your ontology with the two external datasets you identified. Perform useful reasoning over the ontology, and test some interesting queries that make use of reasoning. Finally, start working on the Jupyter notebook (you don’t need to finish it in this milestone or show it to us)."

Grade: 81/100